feat: 完成路由的CRUD

+++ server/doc/ddl.sql
@@ -2,9 +2,9 @@ CREATE TABLE `users` (
     `id` bigint unsigned NOT NULL AUTO_INCREMENT,
     `username` varchar(50) NOT NULL DEFAULT '' COMMENT '用户名',
     `password` varchar(200) NOT NULL DEFAULT '' COMMENT '密码',
-    `created_at` int NOT NULL DEFAULT '0' COMMENT '创建时间',
-    `updated_at` int NOT NULL DEFAULT '0' COMMENT '更新时间',
-    `deleted_at` int NOT NULL DEFAULT '0' COMMENT '删除时间',
+    `created_at` bigint unsigned NOT NULL DEFAULT '0' COMMENT '创建时间',
+    `updated_at` bigint unsigned NOT NULL DEFAULT '0' COMMENT '更新时间',
+    `deleted_at` bigint unsigned NOT NULL DEFAULT '0' COMMENT '删除时间',
     PRIMARY KEY (`id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COMMENT '用户';
 
@@ -18,8 +18,8 @@ CREATE TABLE `routers` (
     `pid` bigint NOT NULL DEFAULT 0 COMMENT '父节点',
     `sort` int NOT NULL DEFAULT 0 COMMENT '排序',
     `component` varchar(255) NOT NULL DEFAULT '' COMMENT '自定义页面',
-    `created_at` int NOT NULL DEFAULT '0' COMMENT '创建时间',
-    `updated_at` int NOT NULL DEFAULT '0' COMMENT '更新时间',
-    `deleted_at` int NOT NULL DEFAULT '0' COMMENT '删除时间',
+    `created_at` bigint unsigned NOT NULL DEFAULT '0' COMMENT '创建时间',
+    `updated_at` bigint unsigned NOT NULL DEFAULT '0' COMMENT '更新时间',
+    `deleted_at` bigint unsigned NOT NULL DEFAULT '0' COMMENT '删除时间',
     PRIMARY KEY (`id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COMMENT '路由';
